requirements.txt
torch>=2.0
torchvision
opencv-python
pyyaml
numpy
Pillow
tqdm
scikit-learn
tensorboard
matplotlib
seaborn
albumentations
segmentation-models-pytorch